requirements.txt
discord.py
chatterbot
chatterbot_corpus
gtts
googletrans